pyproject.toml
[tool.poetry]
name = "seekerbot"
version = "0.1.0"
description = ""
authors = ["Ray <ggammi1007@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
aiogram = "^2.23.1"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
